fix: 코드리뷰 반영 — @Positive shopId/rating BigDecimal/menuRepository findByShopId/테스트 verify 추가 (STORY-12, KAN-110)

+++ src/main/java/com/chunbaetour/domain/shop/dto/response/ShopInfoResponse.java
@@ -2,6 +2,7 @@ package com.chunbaetour.domain.shop.dto.response;
 
 import com.chunbaetour.domain.shop.entity.Menu;
 import com.chunbaetour.domain.shop.entity.Shop;
+import java.math.BigDecimal;
 import java.util.List;
 
 /**
@@ -25,7 +26,7 @@ public record ShopInfoResponse(
         String description,
         String operatingHours,
         String closedDays,
-        double rating,
+        BigDecimal rating,
         int reviewCount,
         boolean isCertified,
         List<MenuResponse> menus
@@ -41,7 +42,7 @@ public record ShopInfoResponse(
                 shop.getDescription(),
                 shop.getOperatingHours(),
                 shop.getClosedDays(),
-                shop.getRating(),
+                BigDecimal.valueOf(shop.getRating()),
                 shop.getReviewCount(),
                 shop.isCertified(),
                 menus.stream().map(MenuResponse::from).toList()
